refactor: rename sqlite bundle
Remove user-facing implementation details from the `SQLite` bundle by
renaming it to `Database`.

References to "SQLite" are kept in tool descriptions and context because
it informs the LLM that their arguments must be valid SQLite.
