+++ database/tool.gpt
@@ -1,13 +1,13 @@
 ---
-Name: SQLite
-Description: Tools for interacting with a SQLite database
+Name: Database
+Description: Tools for interacting with a database
 Metadata: bundle: true
 Share Tools: Query, Exec
 
 ---
 Name: Query
 Description: Run a SQL query against the SQLite database and return the results in markdown format
-Share Context: SQLite Context
+Share Context: Database Context
 Param: query: SQL query to run
 
 #!${GPTSCRIPT_TOOL_DIR}/bin/gptscript-go-tool query
@@ -15,20 +15,20 @@ Param: query: SQL query to run
 ---
 Name: Exec
 Description: Execute a SQL statement against the SQLite database
-Share Context: SQLite Context
+Share Context: Database Context
 Param: statement: SQL statement to execute
 
 #!${GPTSCRIPT_TOOL_DIR}/bin/gptscript-go-tool exec
 
 ---
-Name: SQLite Context
+Name: Database Context
 Type: context
 
 #!${GPTSCRIPT_TOOL_DIR}/bin/gptscript-go-tool context
 
 ---
 !metadata:*:category
-SQLite
+Database
 
 ---
 !metadata:*:icon
